MessageType -> EngineMessageType
Changes
Changed MessageType -> EngineMessageType, opening up the
project to add ProjectMessageType.

Replaced some of the Network interfaces usage of MessageType
with Uint8 so we can keep it generic until we know if it's an
Engine or Project message.

Changed EntityLocator::getEntitiesCoarse(cylinder) to use a
float radius because unsigned int felt unusual.

+++ Source/SharedLib/Simulation/Public/EntityLocator.h
@@ -59,7 +59,7 @@ public:
      * @param radius  The length of the radius to cast.
      */
     std::vector<entt::entity>& getEntitiesCoarse(const Position& cylinderCenter,
-                                                 unsigned int radius);
+                                                 float radius);
 
     /**
      * Overload for BoundingBox.
